backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.9.0
sqlalchemy>=1.4.23
alembic>=1.7.1
pytest>=6.2.5
python-multipart>=0.0.5
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pymupdf>=1.18.0
llama-parse>=0.3.8
nest-asyncio>=1.5.8
